requirements_min.txt
pandas==2.3.2
pydantic==2.11.7
langchain-core==0.3.79
langchain-openai==0.3.35
langchain-tavily==0.2.12
tavily-python==0.7.12
openai==1.104.2
python-dotenv==1.1.1
